Commonize type definitions of Twitter API response
